feat: implement ReportCard component and refactor CompatibilityReportsPage for improved rendering
